feat(all): subscribe methods now return subscription object with dispose method
BREAKING CHANGE - Fixes #10 Subscribe methods now return a Subscription
object which as a dispose method on it for unsubscribing. This matches
the pattern we are using in the new public binding api and we will use
this for any/all consumer facing eventing apis going forward. This also
matches the Rx pattern.